Fix a curl connect timeout issue
We are seeing error like this.
`'WebDriverCurlException' with message 'Curl error thrown for http POST to /session/38719813-dd76-4d78-838c-491016a2f4cf/url with params: {"url":"https:\/\/www.facebook.com\/"}`

This is because the default value for CURLOPT_TIMEOUT_MS is too small. Change it to a reasonable value.