Work around NULL key issue in sendKeys
A long-standing geckodriver bug
(https://bugzilla.mozilla.org/show_bug.cgi?id=1494661) means that it is
not possible to send a key combination such as SHIFT + 'h' + NULL + 'ello'.
In this case the NULL key does not clear the SHIFT modifier as per the
specification, and it additionally prints a character for the NULL key.

It is possible to work around this bug by splitting on the encoded key
and sending the input in multiple commands instead.
Geckodriver already sends an implicity NULL at the end of the sendKeys
command which does work, does release the modifier and does not print
any character.