Work around a bug in Geckodriver (653)
This issue works around a long-standing bug in Geckodriver whereby if
the first child of an anchor is a block-level element, that element
covers the click-point of the anchor. This causes an
ElementNotInteractable exception to be thrown as Marionette believes
that it cannot be scrolled into view.

The workaround to this issue is to attempt to click on an immediate
child node instead. Usually the first child will be sufficient, but if
that first child is a hidden element, then a sibling should instead be
used.

This issue has existed for several years and there is no timeframe for
fixing it upstream.